Added jumping state
