+++ 2D Movement/Packages/ie.itcarlow.2dmovement/Runtime/IdleState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5756c3e0cc7070240aab35aadcaad6f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
