+++ 2D Movement/Packages/ie.itcarlow.2dmovement/Runtime/JumpingState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ef5f07666e56e045a4d358654bbe067
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
