+++ 2D Movement/Packages/ie.itcarlow.2dmovement/Runtime/MovingStateMachine.cs
@@ -7,15 +7,17 @@ public class MovingStateMachine : StateMachine
     [HideInInspector]
     public Runtime2DMovement movement;
     [HideInInspector]
-    public Idle idleState;
-    public MovementLeft movementLeft;
-    public MovementRight movementRight;
+    public IdleState idleState;
+    public WalkingLeftState movementLeft;
+    public WalkingRightState movementRight;
+    public JumpingState jumping;
 
     private void Awake()
     {
-        idleState = new Idle(this);
-        movementLeft = new MovementLeft(this);
-        movementRight = new MovementRight(this);
+        idleState = new IdleState(this);
+        movementLeft = new WalkingLeftState(this);
+        movementRight = new WalkingRightState(this);
+        jumping = new JumpingState(this);
         movement = this.GetComponent<Runtime2DMovement>();
     }
