+++ 2D Movement/Packages/ie.itcarlow.2dmovement/Runtime/WalkingLeftState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b306cd1cec7d9674fb76f175217d9855
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
