+++ 2D Movement/Packages/ie.itcarlow.2dmovement/Runtime/WalkingRightState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c0539e21005ef24ca81a31b17bcdfa4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
